obtention des points + p(X)

+++ TND/calculIntervalle.c
@@ -37,7 +37,7 @@
 
 float calculIntervalle(float moyenne, float variance, int n, float coefficiantAlpha, float *upperLimit)
 {
-	float margeErreur = coefficiantAlpha * sqrt(variance / n);
+	float margeErreur = coefficiantAlpha * (float)sqrt(variance / n);
 	float intervalleInf = moyenne - margeErreur;
 	float intervalleSup = moyenne + margeErreur;
 	*upperLimit = intervalleSup;
